Decompose test support module to comply with 400-line limit
Split the monolithic support.rs (480 lines) into smaller, focused modules:

- fake_daemon.rs (204 lines): FakeDaemon struct, TryCloneStream trait, and
  stream utilities for mock daemon tests
- lifecycle.rs (101 lines): TestLifecycle mock, LifecycleCall recording, and
  health snapshot fixtures
- mod.rs (234 lines): TestWorld, StaticConfigLoader, fixture helpers, and
  re-exports from submodules

This improves code organisation and brings each module under the 400-line
threshold mandated by the project coding guidelines.
